Mrbind wasm trick (#6226)
* Emscripten trick for generating the C bindings.

* Tweak the messages a bit.

* Try adding CI.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try building the bindings in a separate step.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Try again.

* Again.

* Try again.

* Try again.

* Some fixes.

* Update docs for generating the C bindings.

* Adjust emsdk path.

+++ source/MRDotNet2Test/src/BitSetTests.cs
@@ -10,14 +10,14 @@ namespace MRTest
         public void TestDefaultConstructor()
         {
             var a = new BitSet();
-            Assert.That( a.size(), Is.EqualTo(0) );
+            Assert.That( a.size(), Is.EqualTo((nuint)0) );
         }
 
         [Test]
         public void TestConstructor()
         {
             var a = new BitSet( 20 );
-            Assert.That(a.size(), Is.EqualTo(20));
+            Assert.That(a.size(), Is.EqualTo((nuint)20));
         }
 
         [Test]
@@ -34,7 +34,7 @@ namespace MRTest
         {
             var a = new BitSet( 10 );
             a.set( 5 );
-            Assert.That( a.findLast(), Is.EqualTo(5) );
+            Assert.That( a.findLast(), Is.EqualTo((nuint)5) );
         }
 
         [Test]
